move decision_procedure.h to src/solvers
The new API has no dependency on or connection with anything in prop.
