+++ src/solvers/prop/module_dependencies.txt
@@ -1,3 +1,4 @@
+solvers
 solvers/prop
 solvers/bdd
 util
